Lobotomize `nats` into compat-package (#6)
This module now exports a base-4.8 compatible `Numeric.Numeric` module
only for GHC versions prior to GHC 7.9. For GHC 7.9 and later `nats`
becomes an empty package. `Numeric.Numeric.Internal` has been removed,
as well as the `Whole` class and the `natural` function.

This basically folds the code from `Numeric.Natural.Internal`
verbatimely into `Numeric.Natural`.

NB: `nats-1.0` is not yet ready for release. This is just a first
    refactoring step.
