+++ nats.cabal
@@ -1,6 +1,6 @@
 name:          nats
 category:      Numeric, Algebra
-version:       0.2.1
+version:       1.0
 license:       BSD3
 cabal-version: >= 1.10
 license-file:  LICENSE
@@ -37,14 +37,12 @@ flag hashable
 
 library
   default-language: Haskell98
-  hs-source-dirs: src
-  ghc-options: -Wall
 
-  exposed-modules:
-    Numeric.Natural
-    Numeric.Natural.Internal
-
-  build-depends: base >= 2 && < 10
-
-  if flag(hashable)
-    build-depends: hashable >= 1.1 && < 1.3
+  if !impl(ghc>=7.9)
+    -- Starting with GHC 7.10/base-4.8, "Numeric.Natural" lives in `base`
+    hs-source-dirs: src
+    exposed-modules: Numeric.Natural
+    ghc-options: -Wall
+    build-depends: base >= 2 && < 4.8
+    if flag(hashable)
+      build-depends: hashable >= 1.1 && < 1.3
